+++ CMakeLists.txt
@@ -5,4 +5,11 @@ set (CMAKE_CXX_COMPILER "/usr/bin/g++")
 set (CMAKE_CXX_STANDARD 17)
 
 
-add_executable(quadratka main.cpp)
+add_executable(quadratka 
+    main.cpp
+    
+    classes/input/input.cpp
+    classes/calculate/calculate.cpp
+    classes/output/output.cpp
+    classes/structs/structs.cpp
+)
