+++ classes/output/output.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include <assert.h>
+#include <stdio.h>
+
+#include "../structs/structs.h"
+
+void print(const Answer answer);
